requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

python-dotenv==1.0.1
requests==2.32.3
